requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-Login==0.6.3
python-dotenv==1.0.1
PyMySQL==1.1.1
openai==1.66.3
PyMuPDF==1.25.3
requests==2.32.3
email-validator==2.2.0
cryptography
